requirements.txt
Flask==0.10.1
Flask-Login==0.2.9
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.4
Flask-WhooshAlchemy==0.55a
Jinja2==2.7.2
MarkupSafe==0.18
SQLAlchemy==0.9.3
Tempita==0.5.2
WTForms==1.0.5
Werkzeug==0.9.4
Whoosh==2.6.0
blinker==1.3
decorator==3.4.0
gunicorn==18.0
itsdangerous==0.23
pbr==0.6
pocket==0.3.2
psycopg2==2.5.2
requests==2.2.1
sqlalchemy-migrate==0.8.2
wsgiref==0.1.2
